refactor(mode): extract shared surface geometry to src/surface_geom.rs
Move Side / phys_to_pane_local / cell_at_local out of src/tmux/pane_hit.rs into
a neutral src/surface_geom.rs; pane_hit keeps only tmux_pane_at_phys. Repoint
all consumers. No behavior change.
