+++ src/input/hyperlink.rs
@@ -11,7 +11,7 @@
 //! `ozma_terminal`'s shared mouse systems, not here.
 
 use crate::input::{InputPhase, current_modifiers};
-use crate::tmux::pane_hit::{cell_at_local, phys_to_pane_local};
+use crate::surface_geom::{cell_at_local, phys_to_pane_local};
 use crate::webview_pointer::topmost_surface_at;
 use bevy::ecs::entity::Entity;
 use bevy::input::ButtonInput;
